cleanup code, change version numbers. Everything should be release ready!

+++ sirius_gui/src/main/java/de/unijena/bioinf/sirius/gui/fingerid/CompoundStructureImage.java
@@ -136,7 +136,7 @@ class CompoundStructureImage extends JPanel {
 //                gg.drawString(scoreText3, (int) (getWidth() - (tw + w3 + 8)), (int) (getHeight() - 4));
 
             gg.setFont(matchFont);
-            gg.drawString(tanimotoText, (int) (getWidth() - (tw + 4)), (int) (getHeight() - 4));
+            gg.drawString(tanimotoText, (int) (getWidth() - (tw + 4)), getHeight() - 4);
 
 
         }
